fix: rename dev utils file
